Squashed ShaderJobFile refactor. (#59)
The main idea with this PR is to move all assumptions about the locations of files into ShaderJobFileOperations. Existing tests and the WebUi still make assumptions about .png files etc. The rest of the code should only refer to the following (as much as possible):

* shaderJobFile e.g. variant.json (alongside: variant.frag, variant.vert, variant.license, etc.)
* shaderJobResultFile e.g. variant.info.json (alongside: variant.png, variant.txt, etc.)
* ShaderJob: the contents of a shaderJobFile but where the shaders have been parsed into ASTs that can be mutated.
* ImageJob: the contents of a shaderJobFile in memory (Thrift object), but the shaders remain as strings, plus other info on how to run it.
* ImageJobResult: the contents of a shaderJobResultFile in memory (Thrift object), plus extra info.
Using the methods in ShaderJobFileOperations, you can copy, move, delete, check if exists on the files, and convert back and forth between most of the above "types".

We also try to ensure that *all* file operations are implemented in ShaderJobFileOperations.

* Compute shaders: merged it into ImageJob.
* Removed IArtifactManager.
* Simplified IShaderDispatcher.