Rework structs (#66)
A large refactoring to change how structs are handled.

Previously we had pretended that structs were stand-alone declarations, captured their types structurally and used aliases to these types throughout the AST.  This was OK to get things going but (a) made cloning a pain, and (b) was not fully general.

Now we treat structs properly as just part of a variable declaration list.  Instead of aliasing types through the AST we have the notion of a named struct type (which says nothing about fields) and a struct definition  type (which has field information).  A little more work is required to get fields from a named struct type, but the payoff is that cloning is straightforward, and the treatment of structs is a lot more general.