Automate header stripping and make reducer header-aware (#82)
We used to pass around a boolean to say whether a shader was expected to have GraphicsFuzz #defines that would need to be stripped on parsing.  Now this is figured out automatically.

As part of this, #version information is no longer stripped away - it is left and properly parsed. A knock-on effect is that a lot less version information need to be passed around in general, as it is now stored in translation units.

The reducer has been adapted to automatically decide whether to emit headers depending on whether the shaders being reduced contained them.