Tweak comments and docs to help user use adb.
Link to adb documentation without getting too specific. Generally, provide a hint that the device must show in `adb devices`, that the serial number will have a different form if using TCP, and link to the adb documentation if more help is needed.

Also replace "serial ID" with "serial number".