DEV: Rework shader generation in Java (#173)
Shader generation is now written in Java, rather than Python, reducing the extent to which Python is a dependency and making testing of shader generation simpler as it can be done in JUnit.

Fixed an issue where do ... while loops were generated in WebGL 1.0 shaders.

Fixed an infinite loop issue in the generator, where if no available transformation could make progress the transformations would be applied forever.

Added a workaround for a memory exhaustion issue in shader translator.

Added tests for shader family generation.

Changed generation top-level script so that - is used instead of _ to separate arguments, in line with the way glsl-reduce works.

Changed the order of arguments to glsl-generate so that the references are specified before the donors. Removed the need for the user to explicitly make the output directory for shaders.

Replaced --avoid-long-loops with an opposite option, --allow-long-loops, so that long loops are disabled by default.

Replaced --multi-pass with an opposite --single-pass option, so that multi-pass is default.

Made --max-bytes default to 500000, which has been found to work well.

Updated glsl-fuzz walkthrough to reflect these changes.