Fixed issue where attempts to make array access in donated code in-bounds was failing. (#505)
Fixes #503.

The issue was that the AST was being cloned between donating
statements and fixing up the array accesses in said statements, and
the fixup was being attemped on the pre-cloning versions of
statements.  This change (a) adds a test to expose the issue, and (b)
fixes it.

A small amount of cleanup of related code is also included.