Make available builtins aware of shader kind
Now that we more comprehensively support builtins, we have the problem
that not all builtins are available for every kind of shader.  This
change adds a ShaderKind parameter to the function that produces
available builtins, so that only relevant builtins will be returned.
This has required quite a lot of plumbing through, as sometimes
builtins are asked for at relatively deep points in the code, where
the relevant ShaderKind was not always available.

The tests for validity of builtin functions have been extended to try
not just every shading language version and also every kind of shader;
this would catch the (previously uncaught) problem where a fragment
processing builtin could be generated in a compute shader.

Fixes #588.