Avoid pop-up on Windows during testing (#591)
A test uses an empty file as a custom interestingness judge.  Due to not having the .bat extension, this would lead to a pop-up on Windows asking what kind of program should be used to open the file.  As other platforms don't care about the file extension of an executable script, this change uses .bat to avoid the pop-up on Windows.

Fixes #557