Emit GraphicsFuzz defines only when necessary (#595)
This change introduces functionality to check whether a shader uses
one or more of the GraphicsFuzz macros, and only emit the macros if
so. This involved refactoring PrettyPrinterVisitor.emitShader() and
ShaderJobFileOperations.writeShaderJobFile() so that they no longer
take a boolean argument specifying whether the macros should be
written. In the course of doing this refactoring, many places where
null parameters were being passed to the constructor of ReducerContext
(in tests, due to them not being needed) were replaced with non-null,
standard parameters.

The change also gets rid of some never-used 'main' methods that would
otherwise have needed to be fixed up, and fixes a few cases where
shaders were being emitted in non-standard ways.

Fixes #479.