Fix issue where the reducer was replacing a dead code injection with its body
The reducer used to be very conservative about replacing dead code
injections with their bodies and was made less conservative.  However
this introduced a bug whereby the reducer would always replace
'if(_GLF_DEAD(...)) body' by 'body'.  The problem was that the guard
of the dead code injection was being regarded as itself being dead
code, which is wrong; it should only be the 'then' and 'else' branches
of the if statement that are regarded as dead.

This change fixes this problem.  It was necessary to change a number
of tests in the process: tests where the reducer was only managing to
do a particularly good job of reducing a shader due to this bug; for
example it could turn:

if (_GLF_DEAD(...)) {
  int a;
  a = 2;
}

into:

{
  int a;
  a = 2;
}

which is OK, because the body of the 'if' statement does not have
side-effects to data declared outside the 'if' statement.  But this
was only working due to the body of the 'if' being unconditionally
used to replace the 'if'.  Now that that problem is fixed, the side
effect analysis is not yet sophisticated enough to determine that the
body of the 'if' is side-effect free, thus the transformation does not
take place.