Fix opaque expressions based on vector length
We recently added the ability to generate zero as length(zero vector),
and one as length(normalize(one vector)).  However, the code was
incorrectly assuming that the length of a vector of dimension n would
itself be a vector of dimension n, whereas in reality the length of a
vector is always a scalar.

As well as fixing this so that the expected return type for these ways
to make zero and one is 'float', the change allows the dimension of
the vector whose length is taken to be chosen at random.

Finer-grained tests for opaque zero and one generation are also added,
and the shading language versions for which opaque expression
generation is tested has been expanded a bit.

Fixes #559