Replace ArrayInitializer and ScalarInitializer with Initializer
The Initializer class had a never-used subclass, ArrayInitializer
(which was in fact intended to handle the case of initializer lists),
and the ScalarInitializer subclass was being used for all initializers
(including for array initialization).  As initializer lists are not
supported in OpenGL ES, this change simplifies things so that there is
just an Initializer class, which does the job ScalarInitializer was
doing before.

Support for initializer lists can be added in the future if they are
deemed important.

Fixes #520

LocalWords:  ArrayInitializer ScalarInitializer