Reduce switch statements by replacing switch with do..while(false)
Adds a reduction pass that will replace a switch statement with a loop
of the form:

do {
 // Body of switch statement, minus 'case' and 'default' labels
} while (false);

The loop construct is convenient as it allows for 'break' statements.
Existing reduction passes will then try to remove the loop.

Does not apply when --preserve-semantics is used, unless the switch
statement is part of a dead code injection.

Fixes #610