Improve reducer's ability to remove return statements
The reducer needs to be a little conservative about removing non-void
return statements, to avoid making a shader invalid.  This change
makes the reducer less conservative by allowing it to remove a
statement that is or that contains a non-void return statement if
either:

- the enclosing function ends with a non-void return statement that is
  not the statement being considered for removal, or

- the statement being considered for removal is in a block and is
  preceded by a return statement in the block.

Some JavaDoc comments have been added along the way.

Fixes #493