Rework ScopeTreeBuilder
The ScopeTreeBuilder class is one of the most critical classes in the
GraphicsFuzz framework, but had a poor name and was not
well-documented (despite having some subtle behaviours).

This change renames the class to ScopeTrackingVisitor, which is a more
accurate name, and adds documentary comments to its fields and
methods.

The fields have all made private, with accessor and mutator methods
added where needed.

In the process, the ReducerBugPoint and ReducerBugPointBasic classes
have been removed; they used a subclass of ScopeTreeBuilder called
'InformativeScopeTreeBuilder', and rather than maintaining this class
it seemed better to remove those tools (which have not been used for a
long time).