Gracefully handle scenario where no compatible donors are available
It is possible to get into a situation with live or dead code
injection where none of the available donors are compatible with the
reference.  For example, they might all share a named struct in common
(which at present is a source of incompatibility).  When such a
situation arises the generator would throw an exception.  This change
causes the generator, instead, to inject a null statement (;).

Fixes #645.