Use '1;' instead of ';' when replacing jump statements (#724)
Before this change, the null statement, ';', was used when replacing
break, continue, case label and discard statements during live code
injection, and void return statements were completely removed.  The
latter - completely removing a return - caused problems when a return
appeared at the end of a switch statement.  Since it is questionable
whether it is OK for the final case/default label of a switch to be
followed by simply a null statement (e.g. glslangValidator gives a
warning about this), this change makes it so that the statement '1;'
is used in all cases where a null statement was previously used, and
is also used as a replacement for a void return.

Fixes #718.