Prevent the reducer from trying to simplify case labels (#727)
The reducer could throw an exception due to trying to simplify case
statement labels, for which child querying and replacement is not
supported.  Rather than supporting child querying and replacement,
this change prevents the reducer from trying to simplify case labels.
Case labels already need to be literal expressions - so there is not
much to simplify - and changing the values of these literals would be
risky; for example it could easily lead to a switch statement with
duplicate case labels, which is not valid.

Fixes #697.