Enhance reducer to include guard of compound statement when flattening it (#733)
When simplifying a statement of the form:

if(e) {
s1; s2; ...
}

the reducer used to turn it into just:

{ s1; s2; ... }

It now turns it into:

{ e; s1; s2; ... }

For, while and do-while loops are similarly handled so that their
guard (and the initializer and increment components, in the case of a
for loop) are included in the flattened construct that the reducer
produces.

This is useful in cases where e.g. something appearing in the guard is
needed to trigger a bug, but the control flow statement itself is not
needed.

Fixes #732.