Use clamp instead of ternary to make array access in bounds
Live code injection has a feature that attempts to force all array
accesses to be in bounds.  This was being achieved by using a ternary
expression to clamp an array index expression into an acceptable
range, turning:

A[e]

into:

A[e >= 0 && e < len(A) ? e : 0]

A problem with this is that when e has side effects, evaluating e
multiple times can make the access end up being out of bounds anyway.
For example in:

int A[10];
A[++i >= 0 && ++i < 10 ? ++i : 0] = 123;

if i is 8 then ++i will be less than 10, so ++i will be used to index
the array ... but will evaluate to 10.

This change addresses this problem for recent shading language
versions by using the clamp built-in instead of a ternary.  For older
shading language versions we still use a ternary - living with the
problem as we do not deem it worth fixing at present.

The clamping is expressed in a macro, _GLF_MAKE_IN_BOUNDS, to pave the
way for avoiding losing in-bounds clamping during test case reduction.
The reducer has been updated so that in its final simplification pass
it can get rid of these new macros.

Fixes #679.