Fix acosh() input range (#753)
acosh() is only defined for input values >=1, but the shader used values that are less than that (including negative values). This leads to undefined behavior as per SPIR-V spec. In C++ the result may have bee NaN, which is optional in Vulkan.

This change fixes the shader so that all acosh() input values are >=1.