Restrict initializer inlining when preserving semantics (#788)
The reducer tries to be clever and do some initializer inlining when
preserving semantics.  But this can be problematic for a couple of
reasons:

- It can make references and variants hard to compare if initializer
  inlining has been applied to one and not the other

- There are subtle cases where initializer inlining does not preserve
  semantics; for example:

    highp int x = 16777216;
    findLSB(x);

  might not be equivalen to:

    findLSB(16777216);

  since findLSB can operate at low precision when passed a literal
  argument.

This change disables initializer inlining when semantics are being
preserved, unless the initializer occurs in dead code, or is the
initializer for a live-injected variable.

The change contributes some clean-up related to how dead code and
live-injected variables are identified in general.

Partially fixes #639.
Partially fixes #783.