Fix cts test list generator ordering logic
add_amber_test_to_cts.py used amber file names (including .amber suffix)
to order the lines in the cpp file and test names to order the lines in the must pass lists.
This may lead to inconsistent order in cpp and must pass lists in some rare cases.
For example the following test would be ordered in the wrong way if the amber file names are used:
unreachable-discard-statement-in-if.amber
unreachable-discard-statement.amber

This commit changes the cpp file generator to order the lines by test name.