Reducer: simplify literals to canonical form
Before this change, the reducer would not attempt to simplify
literals, so that e.g. 8092.6 would remain in a reduced shader, even
if in dead code.  This was based on the assumption that the exact form
of final literals would not really be important.  However, having
worked a lot with the reducer it is certainly the case that a final
shader that does not involve weird literal values is easier to
understand than one that does.  This change allows the reducer to
simplify any literal that is not in canonical form to one that is.

We should keep an eye on the knock-on effect this has on reducer
performance; by creating more reduction opportunities we could see a
performance drop.