Apply rounding to length of normalized vector
We compute 1.0 as the length of a normalized vector.  Round-off might
mean that we get a result slightly different from 1.0.  This change
introduces a call to 'round', so we compute 1.0 as
'round(length(normalize(non_zero_vector)', to guard against this.  The
use of 'round' means that the transformation is only applicable for
GLSL versions that support 'round'.