Stable version of binary search tree shader
Adds an obviously numerically stable version of the binary search tree
shader.  Eliminates the "simple" version of the shader, which was very
similar to another version of the shader and not obviously numerically
stable.  Avoids using "injectionSwitch" in the stable and original
versions of the binary search tree shader.

Partly fixes #802.