gfauto: skip ignored_crash_signatures
When fuzzing, don't create a report (i.e. discard the test) if the crash signature is in ignored_crash_signatures for the device. Also fix issue revealed by test -- don't create the crash signature directory until needed; this way we can check that the directory was NOT created as part of the test.