gfauto: reduce_source_dir & crash_regex_override (#845)
`gfauto_reduce_source_dir` is a tool that allows you to reduce a source_dir. This can be useful if a reduction failed, or you want to reduce again with a different signature.

`crash_regex_override` will override a crash signature so that nondeterministic bugs can be reduced easily (most likely using `gfauto_reduce_source_dir`). E.g. set `crash_regex_override` to `.*error.*` and the reduction will proceed as long as "error" appears in the log.