Const arrays, and fuzzing of array constructors (#883)
This change does two things:

(1) It corrects the way that qualifiers (in particular const) are
interpreted when applied to array types. A declaration:

const int v[2] = ...;

should be regarded as having type:

QualifiedType(ArrayType(int, 2), CONST)

but was being regarded as having type:

ArrayType(QualifiedType(int, CONST), 2)

(2) It changes the way array constructor expressions are fuzzed.
Previously, an array constructor for an array of size 256 was fuzzed
by trying to make 256 distinct expressions. The chances of the fuzzer
getting stuck and generating a "fuzzed into a corner" exception were
very high in such cases. This change uses an approach with a much
lower probability of failure, whereby a number of element constructor
expressions are generated and then re-used several times at random if
there are not enough to provide a value for every element of the
array.