Support array parameters
This change applies constant folding to the sizes of array parameters
during AST construction, refines equality on array types to regard
array types as equal if and only if their base types are equal and
they agree on constant sizes, and adapts the logic to check whether
arguments match a function prototype to allow array types.

Fixes #784.