Make scopes compulsory when looking for canonical constants (#886)
This change makes the 'scope' parameters of hasCanonicalConstant and
getCanonicalConstant (on Type) compulsory.  An empty scope can be
passed in cases where it would be difficult to get access to a scope
and where it is known that struct name types will not be encountered.

This was motivated by a bug where a scope had been passed in the 'has'
call but not the 'get' call, which this change also fixes.

In the process the change overloads the Scope constructor so that a
global scope (with a null parent) is made from a zero-argument
constructor, rather than by passing in a null parent.  A few other
cleanups suggested by IntelliJ are also applied.