Fixes to typechecker and cleanup (#890)
Some recent changes introduced somewhat duplicated code for unpacking
an array type into a base type and array info object, and the reverse.
This change introduces helper methods in Typer to provide this
functionality, and changes existing code to use those methods.

In addition, it turned out that in a number of places the
"getWithoutQualifiers()" helper method was not being used to get an
unqualified version of this type.  These have been cleaned up, in the
process exposing some test bugs and leading to revamping the checks on
qualifiers that are made during expression generation.