Only use donors of the same shader kind as the reference (#891)
Before this change, only .frag shaders were being used as donors, and
they were being used as donors for .vert and .comp shaders.

This change forces the generator to consider only shaders of the same
kind as the reference for donation into the reference.

Because this allows .comp shaders to be donated into .comp shaders,
which was not happening before, it exposed some limitations relating
to compute shader donation, specifically that the various compute
shader builtin variables were not being handled correctly, and that
interface blocks were not being handled in donation.  A similar issue,
with builtin variables, applied to vertex shaders.

This change introduces solutions to those problems.