Fix to local variable generation during donation (#894)
When generating a local variable for a free variable appearing in a
piece of generated code, we need to strip the variable's type of
qualifiers that don't make sense on local variables (such as 'in',
'out' and 'inout').  A previous change had messed this up.  This adds
tests to expose the problem and fixes the problem.