gfauto: support stable shaders (#899)
With this change, only reference shader jobs that start with "stable_" will lead to wrong image tests. Other tests will just render the variant shader to see if it crashes. This relies on there being shader jobs that are prefixed with "stable_". This change does not bias the random choice of reference towards "stable_" shaders.