Add parentheses when identities are applied to comma expressions (#905)
This PR adds parentheses as needed to comma expressions when they
become arguments to macros or functions, e.g. when the 'e -> min(e,
e)' transformation is applied to the comma expression 'a, b', we want
to end up with 'min((a, b), (a, b))', not 'min(a, b, a, b)'.