Rework expression generation templates
The purpose of this change is principally to add to
VariableIdentifierExprTemplate a new parameter to indicate whether a
variable is a function parameter, because const function parameters
should not be regarded as being compile-time constant.

The changes also introduces JavaDoc comments to the interface for
expression templates, and and cleans up the code of all classes
(including making fields final, removing unecessarily complex use of
generics, and favouring singleton lists over single-element array
lists, for example).