Visit size expressions of ArrayInfo nodes (#915)
We recently added the notion of a size expression to an ArrayInfo
node.  This change enhances StandardVisitor to visit size expressions,
so that they are considered when walking the AST.  Sometimes we don't
want them to be visited, so some visitors have been adapted to
override this method in order to block it.  We also now want to avoid
any sharing of ArrayInfo nodes in the AST, so the change removes its
equals and hashCode methods and adds clonability.