Guard against occurrences of comma in most expressions (#921)
The comma operator should never occur as the top-level argument to an
expression such as a binary or unary expression, with few exceptions
(e.g. a parentheses expression).  This change introduces checks to
guard against this, and to make it clear that using a comma in the
'then' expression of a ternary is not supported (even though it is
allowed in GLSL).