Fix name shadowing problem with switch statements
When injecting switch statements, care was not being taken to account
for variables declared in the original code being added to the scope
used for fuzzing unreachable switch cases that occur after the
reachable code.  This change fixes said problem, including some
reworking of generation parameters that made the fix easier to test.