Smarter removal of loops with limiters (#948)
The statement removal reduction pass is enhanced with an analysis that
figures out those loops whose removal will not affect the limiting
(via loop limiters) of other loops.  This is used when looking for
statements to remove: a live-injected loop whose removal is known not
to impact the limiting of other loops can be removed.