Clean up redundant loop limiter uses in the reducer
This change teaches the reducer when a loop limiter is not actually
used to limit any loop, in which case uses of the loop limiter can be
safely removed or simplified.

This is achieved by extracting existing functionality for analysing
loop limiters into its own class, so that it provides functions that
can be queried from multiple ReductionOpportunities classes.

In the process, static helper functions to test whether variable names
are live variables or loop limiters have been moved into the Constants
helper class, where the prefixes for live variables and loop limiters
are defined.

Fixes #950.