gfauto: fix and update dev dependencies (#1014)
Set `PIPENV_PYTHON` in `dev_shell.sh.template`: In some cases it seems that pipenv will NOT recognize the Python version that is being used, and so pipenv will create a virtual environment using some other version of Python. This line ensures the virtual environment created uses the correct Python binary. This was previously causing Windows CI to fail.

Switch to using the `dataclasses` package, which is a back port of the Python 3.7 dataclasses feature. We were previously using `attr` package.

Re-generate Pipfile.lock so that we get the dataclasses package and updated versions of all other dependencies. 

Fix new lint errors.

Update instructions.