gfauto: change reduction output
Rename the setting "skip_spirv_reduce" -> "skip_semantics_changing_reduction" and make sure it applies to GLSL reductions as well.

Change the summary output directory for a test so that we get "unreduced", "reduced_1", "reduced_2" (if semantics changing reduction was done), and their corresponding result directories with the suffix "_result". For the "reduced_1" directory, if the "reference" shader directory was renamed to "_reference" (because this is a crash bug so the reference is no longer needed) AND the shader was derived from a "stable_" shader then rename "_reference" back to "reference"; this allows generating a "wrong image" test from this directory, even though the bug was a crash bug.

The reduction result directory for spirv-fuzz now gets symlinks "1" and "2" that point to the semantics-preserving and semantics-changing reduction result directories, respectively, mirroring the GLSL reduction result directory.

Fixed tests due to the renamed summary directories.