gfauto: generate variant_2
When creating the summary of a crash bug derived from a `stable_` shader, `reduction_1/` will contain a "wrong image" test that exposes the bug. In this case, we now add a `variant_2` shader that is the same as the variant shader, but with the last transformation removed. This is likely to give a shader that is very similar to the variant but does not trigger the bug. The `variant_2` shader is created using the `--replay-range` option of spirv-fuzz.