Do not add bindings LiteralToUniformReductionOpportunity
The LiteralToUniformReductionOpportunity adds new uniforms to a shader
job.  It should not provide these uniforms with bindings -- it is up
to the reducer to add bindings if they are deemed necessary, right
before emitting the shader as a text file.

This change fixes that issue.  However, it turns out that the reducer
was not adding bindings in the case that no uniforms were present in
the first place -- it was treating a uniform-free shader as a
non-Vulkan shader.  As a temporary workaround, a uniform-free shader
is now treated by the reducer as a Vulkan shader so that if uniforms
are then created, they get bindings when written out as text files.

Related issue to fix this more elegantly in the long-run: #1046.