Provide uniform arguments as strings
PipelineInfo provides a method for getting the numeric arguments for a
uniform.  This previously returned a list of type Number.  It turned
out that the JSON library backing PipelineInfo might sometimes
represent numeric values using non-standard subclasses of Number - in
particular LazilyParsedNumber.

To guard against this, numeric arguments are now provided as strings.
We always know the types of uniform variables, so can interpret these
strings as integers or floats accordingly.

Fixes #1054.